msm8998-common: Kang walleye offsets for GPU composition
Window animations now have rounded corners, meaning that
the shader that we use is more expensive than it used to be.

Because of this, we need to adjust the SF offsets, to make sure
that we'll have enough time to finish the GPU composition without
dropping frames.

These offsets as similar to blueline's, adjusted slightly because
we were getting false "frame missed" signals if composition
starts too early.

Change-Id: I728e40adc0ed46c54c610fc7e7cb83627f00ef1a
Signed-off-by: NurKeinNeid <mralexman3000@gmail.com>